backend/requirements_clean.txt
setuptools
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models
boto3==1.40.40
fastapi==0.117.1
face-recognition==1.3.0
face_recognition_models==0.3.0
google-ai-generativelanguage==0.6.15
google-api-python-client==2.186.0
google-generativeai==0.8.5
langchain-community==0.3.30
livekit==1.0.13
livekit-agents==1.2.11
livekit-plugins-google==1.2.11
livekit-plugins-noise-cancellation==0.2.5
livekit-protocol==1.0.6
nest-asyncio==1.6.0
numpy==2.2.6
pandas==2.3.2
PyPDF2==3.0.1
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
SpeechRecognition==3.10.4
uvicorn==0.37.0
httpx==0.28.1
duckduckgo-search==8.1.1
dlib==20.0.0
opencv-python==4.10.0.84
mem0ai
